revert change to debugging statement that shouldn't be there yet
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1858 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/winio.c
@@ -3211,7 +3211,7 @@ void dump_buffer_reverse(void)
     const filestruct *fileptr = filebot;
 
     while (fileptr != NULL) {
-	fprintf(stderr, "(%lu) %s\n", (unsigned long)fileptr->lineno, fileptr->data);
+	fprintf(stderr, "(%d) %s\n", fileptr->lineno, fileptr->data);
 	fileptr = fileptr->prev;
     }
 }
